requirements.txt
asgiref==3.5.2
cmake==3.24.1.1
contourpy==1.0.5
cycler==0.11.0
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8==5.0.4
fonttools==4.37.2
kiwisolver==1.4.4
matplotlib==3.5.0
mccabe==0.7.0
numpy==1.23.3
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
seaborn==0.12.0
setuptools-scm==7.0.5
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
